requirements.txt
appdirs==1.4.4
arrow==1.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
flask==3.0.3
flexcache==0.3
flexparser==0.3.1
h11==0.14.0
idna==3.7
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
numpy==2.0.0
oauthlib==3.2.2
outcome==1.3.0.post0
pandas==2.2.2
Pint==0.24.1
pydantic==1.10.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stravalib==1.6
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
